requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Bcrypt
Flask-Migrate
flask-paginate
WTForms
email-validator
Werkzeug
python-dotenv
reportlab
gunicorn
livereload
schedule
requests
cryptography
psycopg2-binary
